Adding VideoGame Review Page
there are some problems, that are currently being worked out. You are able to add a game onto the server and give the title. I want to have a DB of game titles and it will check the name and if its a match provided the image. I want users to be able to leave reviews and comments and a rating system and save it all in a DB table. 

Problems: The plus sign doesnt move over when a new title is created, can only have 4 titles per row.